LayoutTests/fast/js/typedarray-set-overlapping-elements-of-same-size-expected.txt

Tests the code path of typedArray.set that tries to do a memmove-with-conversion for overlapping arrays.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo(10) is [42,42,42,42,42,42,42,42,42,42,42]
PASS bar(10) is [42,43,44,45,46,47,48,49,50,51,51]
PASS successfullyParsed is true

TEST COMPLETE
